Simply updates

+++ routes/mytasks.js
@@ -31,18 +31,18 @@ router.post('/', auth, async (req, res) => {
 router.put('/:id',[auth,validateObjectId], async (req, res) => {
   const { error } = validate(req.body); 
   if (error) return res.status(400).send(error.details[0].message);
+  const check = await Group
+    .findById(req.params.id);
 
-  const old = await Task.findById(req.params.id);
-  console.log(old);
-  
+  if (req.user._id !== check.owner) return res.status(403).send('Access denied.');
+
   const task = await Task.findByIdAndUpdate(req.params.id, {
     
         title: req.body.title,
         description: req.body.description,
         status: req.body.status,
-        tag: req.body.tag,
-      },
-      {  new: true,}
+        tags: req.body.tags,
+      }
       );
 
   if (!task) return res.status(404).send('The task with the given ID was not found.');
@@ -50,6 +50,10 @@ router.put('/:id',[auth,validateObjectId], async (req, res) => {
 });
 
 router.delete('/:id', [auth, validateObjectId], async (req, res) => {
+  const check = await Group
+    .findById(req.params.id);
+
+  if (req.user._id !== check.owner) return res.status(403).send('Access denied.');
   const task = await Task.findByIdAndRemove(req.params.id);
 
   if (!task) return res.status(404).send('The task with the given ID was not found.');
@@ -57,7 +61,7 @@ router.delete('/:id', [auth, validateObjectId], async (req, res) => {
   res.send(task);
 });
 
-router.get('/:id', async (req, res) => {
+router.get('/:id', [auth, validateObjectId], async (req, res) => {
   const task = await Task.findById(req.params.id);
 
   if (!task) return res.status(404).send('The task with the given ID was not found.');
